refactor: split menu screen
