+++ src/modules/menu.h
@@ -1,0 +1,11 @@
+#ifndef MENU_H
+#define MENU_H
+
+#include <M5GFX.h>
+
+#include "display.h"  // mainCanvas を使用するため
+
+// メニュー画面を描画する関数
+void drawMenuScreen();
+
+#endif  // MENU_H
